win build fixed III

+++ libshv/tests/libshvcore/chainpack/rpcvalue/tst_chainpack_rpcvalue.cpp
@@ -236,8 +236,8 @@ CHAINPACK_TEST_CASE(text_test) {
 
 	class Point {
 	public:
-		long x;
-		long y;
+		int x;
+		int y;
 		Point (int x, int y) : x(x), y(y) {}
 		RpcValue to_json() const { return RpcValue::List { x, y }; }
 	};
